HuggingFaceAgentsCourse/agentIntro/pyproject.toml
[project]
name = "agentintro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.72.0",
    "python-dotenv>=1.1.0",
]
